Create seperate classes for interfaces and enums
