+++ firebase-app-distribution/src/main/java/com/google/firebase/appdistribution/FirebaseAppDistributionRegistrar.java
@@ -15,13 +15,11 @@
 package com.google.firebase.appdistribution;
 
 import androidx.annotation.Keep;
-
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.components.Component;
 import com.google.firebase.components.ComponentRegistrar;
 import com.google.firebase.components.Dependency;
 import com.google.firebase.platforminfo.LibraryVersionComponent;
-
 import java.util.Arrays;
 import java.util.List;
 
@@ -31,14 +29,10 @@ public class FirebaseAppDistributionRegistrar implements ComponentRegistrar {
   @Override
   public List<Component<?>> getComponents() {
     return Arrays.asList(
-            Component.builder(FirebaseAppDistribution.class)
-                    .add(Dependency.required(FirebaseApp.class))
-                    .factory(c -> new FirebaseAppDistribution(c.get(FirebaseApp.class)))
-                    .build(),
-            LibraryVersionComponent.create("fire-app-distribution", BuildConfig.VERSION_NAME));
+        Component.builder(FirebaseAppDistribution.class)
+            .add(Dependency.required(FirebaseApp.class))
+            .factory(c -> new FirebaseAppDistribution(c.get(FirebaseApp.class)))
+            .build(),
+        LibraryVersionComponent.create("fire-app-distribution", BuildConfig.VERSION_NAME));
   }
-
-
-
-
 }
